+++ src/main/java/kitchenpos/menu/repository/dao/JdbcTemplateMenuProductDao.java
@@ -1,11 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.menu.repository.dao;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.repository.dao.MenuProductDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
