+++ src/main/java/kitchenpos/menu/repository/dao/MenuDao.java
@@ -1,8 +1,8 @@
-package kitchenpos.dao;
+package kitchenpos.menu.repository.dao;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 public interface MenuDao {
     Menu save(Menu entity);
