+++ src/test/java/kitchenpos/menu/application/MenuGroupServiceTest.java
@@ -1,11 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import kitchenpos.application.dto.request.MenuGroupCreateRequest;
-import kitchenpos.application.dto.response.MenuGroupResponse;
+import kitchenpos.application.ServiceTest;
+import kitchenpos.menu.ui.request.MenuGroupCreateRequest;
+import kitchenpos.menu.response.MenuGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
